+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-##Team member: Castiel Li (CL) , Artur G (AG) , Nate C(NC) , Miranda S (MS)
+##Team member: Castiel Li (CL) , Artur G (AG) , Nate C(NC) , Miranda P (MP)
 
 ## [0.0.1]
 ### 2018-03-10
@@ -10,6 +10,11 @@ All notable changes to this project will be documented in this file.
 * Added timepicker.js component -CL
 * Added Changelog -CL
 * Added Readme -CL
+* Added statusbar.js component -MP
+* Added actionbutton.js component -MP
+* Added listtaks.js component -MP
+* Added taskdetail.js component -MP
+* Added firebase.js component -MP
 
 ### Updated
 * updated timepicker.js (display 10 mins time block instead of mins) -CL